modules/clustering-impl/src/main/java/org/clueminer/clustering/aggl/Element.java
/*
 * Copyright (C) 2011-2018 clueminer.org
 *
 * This program is free software: you can redistribute it and/or modify
 * it under the terms of the GNU General Public License as published by
 * the Free Software Foundation, either version 3 of the License, or
 * (at your option) any later version.
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU General Public License for more details.
 *
 * You should have received a copy of the GNU General Public License
 * along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
package org.clueminer.clustering.aggl;

/**
 *
 * @author Tomas Barton
 */
public class Element implements Comparable {

    private double value;
    private int row;
    private int column;

    public Element(double value, int row, int column) {
        this.value = value;
        this.row = row;
        this.column = column;
    }

    @Override
    public int compareTo(Object o) {
        Element other = (Element) o;
        double diff = this.value - other.value;
        if (diff > 0) {
            return 1;
        } else if (diff < 0) {
            return -1;
        }
        return 0;
    }

    public double getValue() {
        return value;
    }

    public void setValue(double value) {
        this.value = value;
    }

    public int getRow() {
        return row;
    }

    public void setRow(int row) {
        this.row = row;
    }

    public int getColumn() {
        return column;
    }

    public void setColumn(int column) {
        this.column = column;
    }

    @Override
    public String toString() {
        StringBuilder sb = new StringBuilder("Element [ ");
        sb.append(String.format("%.2f", value)).append(" (").append(row).append(", ").append(column);
        sb.append(")]");
        return sb.toString();
    }

}
